CG-3
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@5055 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/BinaryGrammar.cpp
@@ -57,7 +57,7 @@ int BinaryGrammar::parse_grammar_from_file(const char *filename, const char *, c
 		CG3Quit(1);
 	}
 	else {
-		grammar->grammar_size = (uint32_t)_stat.st_size;
+		grammar->grammar_size = static_cast<size_t>(_stat.st_size);
 	}
 
 	FILE *input = fopen(filename, "rb");
